+++ test/support/data_case.ex
@@ -35,7 +35,7 @@ defmodule TdLm.DataCase do
       Sandbox.mode(TdLm.Repo, {:shared, self()})
       parent = self()
 
-      case Process.whereis(TdLm.RelationLoader) do
+      case Process.whereis(TdLm.Cache.LinkLoader) do
         nil -> nil
         pid -> Sandbox.allow(TdLm.Repo, parent, pid)
       end